feat: generate trip log odometer readings

+++ app/Models/User.php
@@ -38,7 +38,7 @@ class User extends Authenticatable
         return [
             'email_verified_at' => 'datetime',
             'password' => 'hashed',
-            'car_mileage' => 'integer',
+            'car_mileage' => 'float',
         ];
     }
